# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API REST para reservas, servicios, combos y administracion de Atrevida Fit.",
        "title": "Atrevida Fit - Agenda API",
        "contact": {},
        "version": "v0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/bd/servicios": {
            "get": {
                "description": "Devuelve servicios desde PostgreSQL con filtros. Filtros: nombre busqueda parcial (opcional), categoria busqueda parcial (opcional), local SAN MARTIN/PASEO ARANJUEZ (opcional), sesiones numero exacto (opcional), requiere_evaluacion true/false (opcional). Response: total (int), filtros (objeto con nombre, categoria, local, sesiones, requiere_evaluacion), servicios ([]ServicioItem con: id, nombre, categoria, local, tiempo HH:MM, costo, sesiones, tipoEspacio M/B, requiere_evaluacion).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios BD"
                ],
                "summary": "Listar servicios desde base de datos",
                "parameters": [
                    {
                        "type": "string",
                        "example": "depila",
                        "description": "Busqueda parcial por nombre",
                        "name": "nombre",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "Corporal",
                        "description": "Busqueda parcial por categoria",
                        "name": "categoria",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "SAN MARTIN",
                            "PASEO ARANJUEZ"
                        ],
                        "type": "string",
                        "example": "SAN MARTIN",
                        "description": "Local",
                        "name": "local",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 6,
                        "description": "Numero exacto de sesiones",
                        "name": "sesiones",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "Filtrar por servicios que requieren evaluacion",
                        "name": "requiere_evaluacion",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.servicioListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: sesiones debe ser entero positivo, local invalido, requiere_evaluacion debe ser true/false",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un servicio en PostgreSQL y opcionalmente lo asocia a un local. Body: nombre (requerido), categoria (requerido), tiempo HH:MM (opcional), costo (opcional), sesiones entero positivo default 1 (opcional), tipo_espacio_requerido M/B (opcional), requiere_evaluacion true/false default true (opcional), local para activar (opcional). Response: id (int ID del servicio creado).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios BD"
                ],
                "summary": "Crear servicio",
                "parameters": [
                    {
                        "description": "Datos del servicio",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.crearServicioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.idResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: nombre/categoria requerido, tipo_espacio_requerido invalido, local no encontrado, local sin espacios",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/bd/servicios/local/{id}": {
            "post": {
                "description": "Asocia un servicio existente a un local. Param: id del servicio (requerido, path). Body: local (requerido). Response: mensaje string.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios BD"
                ],
                "summary": "Activar servicio en local",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 8,
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Local donde activar el servicio",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.activarServicioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido, local requerido, servicio/local no encontrado, local sin espacios",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/bd/servicios/{id}": {
            "get": {
                "description": "Devuelve un servicio por su ID. Param: id (requerido, path). Response: servicio (ServicioItem con: id, nombre, categoria, local, tiempo HH:MM, costo, sesiones, tipoEspacio M/B, requiere_evaluacion).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios BD"
                ],
                "summary": "Obtener servicio por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 8,
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.servicioItemResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Realiza borrado logico de un servicio (activo=false). Param: id (requerido, path). Response: mensaje string.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios BD"
                ],
                "summary": "Eliminar servicio",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 8,
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado o ya inactivo",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Actualiza un servicio existente. Solo se actualizan los campos enviados. Param: id (requerido, path). Body: nombre (opcional), categoria (opcional), tiempo HH:MM (opcional), costo (opcional), sesiones (opcional), tipo_espacio_requerido M/B (opcional), requiere_evaluacion true/false (opcional), activo true/false (opcional). Response: mensaje string.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Servicios BD"
                ],
                "summary": "Actualizar servicio",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 8,
                        "description": "ID del servicio",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Campos a actualizar (todos opcionales, al menos uno requerido)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.actualizarServicioRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.messageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Error de validacion: id invalido, tipo_espacio_requerido invalido, sin campos a modificar",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Servicio no encontrado",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.activarServicioRequest": {
            "type": "object",
            "required": [
                "local"
            ],
            "properties": {
                "local": {
                    "type": "string",
                    "example": "PASEO ARANJUEZ"
                }
            }
        },
...
        "handlers.actualizarServicioRequest": {
            "type": "object",
            "properties": {
                "activo": {
                    "type": "boolean",
                    "example": true
                },
                "categoria": {
                    "type": "string",
                    "example": "Corporal"
                },
                "costo": {
                    "type": "number",
                    "example": 420
                },
                "nombre": {
                    "type": "string",
                    "example": "Depilacion Laser Piernas Premium"
                },
                "requiere_evaluacion": {
                    "type": "boolean",
                    "example": false
                },
                "sesiones": {
                    "type": "integer",
                    "example": 8
                },
                "tiempo": {
                    "type": "string",
                    "example": "01:15"
                },
                "tipo_espacio_requerido": {
                    "type": "string",
                    "example": "M"
                }
            }
        },
...
        "handlers.crearServicioRequest": {
            "type": "object",
            "required": [
                "categoria",
                "nombre"
            ],
            "properties": {
                "categoria": {
                    "type": "string",
                    "example": "Corporal"
                },
                "costo": {
                    "type": "number",
                    "example": 350
                },
                "local": {
                    "description": "opcional",
                    "type": "string",
                    "example": "SAN MARTIN"
                },
                "nombre": {
                    "type": "string",
                    "example": "Depilacion Laser Piernas"
                },
                "requiere_evaluacion": {
                    "type": "boolean",
                    "example": true
                },
                "sesiones": {
                    "type": "integer",
                    "example": 6
                },
                "tiempo": {
                    "type": "string",
                    "example": "01:00"
                },
                "tipo_espacio_requerido": {
                    "description": "\"M\" | \"B\" | nil",
                    "type": "string",
                    "example": "M"
                }
            }
        },
...
        "handlers.idResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
        "handlers.messageResponse": {
            "type": "object",
            "properties": {
                "mensaje": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                }
            }
        },
...
        "handlers.servicioItemResponse": {
            "type": "object",
            "properties": {
                "servicio": {
                    "$ref": "#/definitions/models.ServicioItem"
                }
            }
        },
        "handlers.servicioListResponse": {
            "type": "object",
            "properties": {
                "filtros": {
                    "$ref": "#/definitions/handlers.servicioFiltrosResponse"
                },
                "servicios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ServicioItem"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
        "utils.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "operacion realizada correctamente"
                },
                "status": {
                    "type": "string",
                    "example": "OK"
                }
            }
        }
    }
}